requirements.txt
pymongo[srv]==4.11
selenium==4.28.1
tiktoken==0.8.0
python-dotenv==1.0.1
langchain==0.3.17
langchain_core==0.3.33
langchain_groq==0.2.4